reorganize pep508 marker module
